Resolve inadvertent dependency upgrade
The 3.x release series of this library is meant to coincide
with cspray/annotated-container v2.x. That version of AC relies
on cspray/annotated-target ^0.3. Versions 3.2 and 3.3 of this
library broke backwards compatibility by relying on a version
of annotated-target that is not compatible with the AC 2.x
series.

This patch downgrades 3.4 to use cspray/architectural-decision 1.4,
that still uses a version of annotated-target compatible with AC.